feat: make sure we do not accidently delete serviceaccount

+++ pkg/services/sqlstore/user.go
@@ -734,7 +734,12 @@ func deleteUserInTransaction(sess *DBSession, cmd *models.DeleteUserCommand) err
 	if !has {
 		return models.ErrUserNotFound
 	}
-
+	// we need to make sure that usercommand
+	// do not accidently delete
+	// a service account when the intention was a user account
+	if user.IsServiceAccount && !cmd.DeleteServiceAccount {
+		return fmt.Errorf("user is service account, but cmd said it should not delete service account")
+	}
 	deletes := []string{
 		"DELETE FROM star WHERE user_id = ?",
 		"DELETE FROM " + dialect.Quote("user") + " WHERE id = ?",
